frontend progress; creating backend functionality for Recent Activity

+++ backend/content-service/src/modules/content/content.controller.ts
@@ -18,35 +18,96 @@ import { Serializer } from '../../serializer';
 import { DataSourceEnumDTO } from '../../proto/common';
 import { Context } from '../../class/Context';
 import { SearchQuery } from '../../class/search/SearchQuery';
+import { ActivityService } from '../activity/activity.service';
+import { ActivityRecord } from '../activity/entities/ActivityRecord';
 
 @Controller()
 export class ContentController {
     constructor(
-        @Inject('IContentService') private readonly contentService: ContentService<ContentBase>
+        @Inject('IContentService') private readonly contentService: ContentService<ContentBase>,
+        @Inject('IActivityService') private readonly activityService: ActivityService,
     ) { }
 
     @GrpcMethod('ContentService', 'create')
     async create(request: CreateContentRequest): Promise<CreateContentResponse> {
-        const { entityName, contentBody, source } = request;
+        const { entityName, contentBody, source, context } = request;
+        const worldId = context?.world
+        const userId = context?.user
+        if (!worldId) throw new Error(`worldId cannot be undefined`)
+        if (!userId) throw new Error(`userId cannot be undefined`)
         if (!contentBody) throw new Error(`contentBody cannot be undefined`)
         const entity = deserializeContentBodyDTO(contentBody)
-        return this.contentService.create(entityName, entity, deserializeEnum(DataSourceEnumDTO, DataSourceEnum, source));
+        const result = await this.contentService.create(entityName, entity, deserializeEnum(DataSourceEnumDTO, DataSourceEnum, source));
+
+        // record activity
+        const newActivityRecord: ActivityRecord = ActivityRecord.build({
+            eventName: "CONTENT_CREATED",
+            label: `Created content`,
+            relatedTargetEntity: entityName,
+            relatedTargetId: result.id,
+            world: { id: worldId },
+            user: { id: userId }
+        })
+        await this.activityService.create(
+            newActivityRecord,
+            userId,
+            worldId,
+            deserializeEnum(DataSourceEnumDTO, DataSourceEnum, source)
+        )
+
+        return { contentBody: Serializer.toDTO(result) };
     }
 
     @GrpcMethod('ContentService', 'update')
     async update(request: UpdateContentRequest): Promise<UpdateContentResponse> {
-        const { entityName, contentBody, context } = request;
+        const { entityName, contentBody, context, source } = request;
+        const worldId = context?.world
+        const userId = context?.user
+        if (!worldId) throw new Error(`worldId cannot be undefined`)
+        if (!userId) throw new Error(`userId cannot be undefined`)
         if (!contentBody) throw new Error(`contentBody cannot be undefined`)
-        if (!context) throw new Error(`context cannot be undefined`)
         const entity = deserializeContentBodyDTO(contentBody)
-        return this.contentService.update(entityName, entity, Context.fromDTO(context));
+
+        const result = await this.contentService.update(entityName, entity, Context.fromDTO(context));
+
+        // record activity
+        const newActivityRecord: ActivityRecord = ActivityRecord.build({
+            eventName: "CONTENT_UPDATED",
+            label: `Updated content`,
+            relatedTargetEntity: entityName,
+            relatedTargetId: result.id,
+            world: { id: worldId },
+            user: { id: userId }
+        })
+        await this.activityService.create(
+            newActivityRecord,
+            userId,
+            worldId,
+            deserializeEnum(DataSourceEnumDTO, DataSourceEnum, source)
+        )
+
+        return { contentBody: Serializer.toDTO(result) };
     }
 
     @GrpcMethod('ContentService', 'delete')
     async delete(request: DeleteContentRequest): Promise<DeleteContentResponse> {
-        const { entityName, id, context } = request;
-        if (!context) throw new Error(`context cannot be undefined`)
+        const { entityName, id, context, source } = request;
+        const worldId = context?.world
+        const userId = context?.user
+        if (!worldId) throw new Error(`worldId cannot be undefined`)
+        if (!userId) throw new Error(`userId cannot be undefined`)
         await this.contentService.delete(entityName, id, Context.fromDTO(context));
+
+        // record activity
+        const newActivityRecord: ActivityRecord = ActivityRecord.build({
+            eventName: "CONTENT_DELETED",
+            label: `Deleted content`,
+            relatedTargetEntity: entityName,
+            relatedTargetId: id,
+            world: { id: worldId },
+            user: { id: userId }
+        })
+        await this.activityService.create(newActivityRecord, userId, worldId, deserializeEnum(DataSourceEnumDTO, DataSourceEnum, source));
         return { success: true };
     }
 
@@ -63,6 +124,12 @@ export class ContentController {
         const { requests } = request;
         const entityName = requests[0]?.entityName;
         const source = requests[0]?.source;
+        const context = requests[0]?.context;
+        if (!context) throw new Error(`context cannot be undefined`)
+        const worldId = context?.world
+        const userId = context?.user
+        if (!worldId) throw new Error(`worldId cannot be undefined`)
+        if (!userId) throw new Error(`userId cannot be undefined`)
         const entities = requests.map(req => {
             if (!req.contentBody) throw new Error(`contentBody cannot be undefined`);
             return deserializeContentBodyDTO(req.contentBody);
@@ -70,6 +137,16 @@ export class ContentController {
 
         const results = await this.contentService.createBulk(entityName, entities, deserializeEnum(DataSourceEnumDTO, DataSourceEnum, source));
         const responses: CreateContentResponse[] = results.map(result => ({ contentBody: Serializer.toDTO(result) }));
+
+        // record activity
+        const newActivityRecord: ActivityRecord = ActivityRecord.build({
+            eventName: "CONTENT_CREATED_BULK",
+            label: `Created ${entities.length} content in bulk`,
+            relatedTargetEntity: entityName,
+            world: { id: worldId },
+            user: { id: userId }
+        })
+        await this.activityService.create(newActivityRecord, userId, worldId, deserializeEnum(DataSourceEnumDTO, DataSourceEnum, source));
         return { responses };
     }
 
@@ -77,8 +154,13 @@ export class ContentController {
     async updateBulk(request: UpdateBulkContentRequest): Promise<UpdateBulkContentResponse> {
         const { requests } = request;
         const entityName = requests[0]?.entityName;
+        const source = requests[0]?.source;
         const context = requests[0]?.context;
         if (!context) throw new Error(`context cannot be undefined`)
+        const worldId = context?.world
+        const userId = context?.user
+        if (!worldId) throw new Error(`worldId cannot be undefined`)
+        if (!userId) throw new Error(`userId cannot be undefined`)
         const entities = requests.map(req => {
             if (!req.contentBody) throw new Error(`contentBody cannot be undefined`);
             return deserializeContentBodyDTO(req.contentBody);
@@ -86,6 +168,16 @@ export class ContentController {
 
         const results = await this.contentService.updateBulk(entityName, entities, Context.fromDTO(context));
         const responses: UpdateContentResponse[] = results.map(result => ({ contentBody: Serializer.toDTO(result) }));
+
+        // record activity
+        const newActivityRecord: ActivityRecord = ActivityRecord.build({
+            eventName: "CONTENT_UPDATED_BULK",
+            label: `Updated ${entities.length} content in bulk`,
+            relatedTargetEntity: entityName,
+            world: { id: worldId },
+            user: { id: userId }
+        })
+        await this.activityService.create(newActivityRecord, userId, worldId, deserializeEnum(DataSourceEnumDTO, DataSourceEnum, source));
         return { responses };
     }
 
@@ -94,10 +186,25 @@ export class ContentController {
         const { requests } = request;
         const entityName = requests[0]?.entityName; // Assuming all requests are for same entity type
         const ids = requests.map(req => req.id);
-        const context = requests[0]?.context;
+        const context = requests[0]?.context; // Assuming all requests are for same context
+        const source = requests[0]?.source; // Assuming all requests are for same source
         if (!context) throw new Error(`context cannot be undefined`)
+        const worldId: string | undefined = context?.world
+        const userId: string | undefined = context?.user
+        if (!worldId) throw new Error(`worldId cannot be undefined`)
+        if (!userId) throw new Error(`userId cannot be undefined`)
         await this.contentService.deleteBulk(entityName, ids, Context.fromDTO(context));
         const responses = requests.map(() => ({ success: true }));
+
+        // record activity
+        const newActivityRecord: ActivityRecord = ActivityRecord.build({
+            eventName: "CONTENT_DELETED_BULK",
+            label: `Deleted ${ids.length} content in bulk`,
+            relatedTargetEntity: entityName,
+            world: { id: worldId },
+            user: { id: userId }
+        })
+        await this.activityService.create(newActivityRecord, userId, worldId, deserializeEnum(DataSourceEnumDTO, DataSourceEnum, source));
         return { responses };
     }
     @GrpcMethod('ContentService', 'getStats')
